requirements.txt
pandas
sqlalchemy
psycopg2-binary
scikit-learn
PuLP
streamlit
plotly
tqdm
numpy
python-dotenv
